1.拓展了spring的自动注入机制, 添加了一个静态注入处理器,可以通过上下文自动注入到静态字段/方法 2. 修改了AofFile的实现, 增加了flusher 3. 更改了作者名

+++ src/main/java/org/hashdb/ms/exception/DBClientException.java
@@ -8,11 +8,11 @@ import org.hashdb.ms.net.nio.msg.v1.Message;
  * Date: 2023/11/21 12:47
  * DB 外部的错误，是客户端不当操作引发的错误
  *
- * @author huanyuMake-pecdle
+ * @author Huanyu Mark
  */
 @StandardException
 public class DBClientException extends RuntimeException {
-    public ErrorMessage msg(long actId) {
+    public ErrorMessage msg(int actId) {
         return new ErrorMessage(actId, this);
     }
