+++ src/main/java/org/hashdb/ms/net/nio/msg/v1/ActAppCommandMessage.java
@@ -3,14 +3,14 @@ package org.hashdb.ms.net.nio.msg.v1;
 /**
  * Date: 2024/1/18 21:54
  *
- * @author huanyuMake-pecdle
+ * @author Huanyu Mark
  */
 public class ActAppCommandMessage extends ActMessage<Object> {
-    public ActAppCommandMessage(long id, long actId, Object body) {
+    public ActAppCommandMessage(int id, int actId, Object body) {
         super(id, actId, body);
     }
 
-    public ActAppCommandMessage(long actId, Object body) {
+    public ActAppCommandMessage(int actId, Object body) {
         super(actId, body);
     }
