+++ src/main/java/org/hashdb/ms/net/nio/msg/v1/ActMessage.java
@@ -3,23 +3,23 @@ package org.hashdb.ms.net.nio.msg.v1;
 /**
  * Date: 2024/1/17 16:35
  *
- * @author huanyuMake-pecdle
+ * @author Huanyu Mark
  */
 public abstract class ActMessage<B> extends Message<B> {
 
-    protected final long actId;
+    protected final int actId;
 
-    public ActMessage(long id, long actId, B body) {
+    public ActMessage(int id, int actId, B body) {
         super(id, body);
         this.actId = actId;
     }
 
-    public ActMessage(long id, Message<?> request, B body) {
+    public ActMessage(int id, Message<?> request, B body) {
         super(id, body);
         this.actId = request.id;
     }
 
-    public ActMessage(long actId, B body) {
+    public ActMessage(int actId, B body) {
         super(body);
         this.actId = actId;
     }
@@ -29,11 +29,11 @@ public abstract class ActMessage<B> extends Message<B> {
         this.actId = request.id;
     }
 
-    public static ActMessage<String> act(long actId) {
+    public static ActMessage<String> act(int actId) {
         return new DefaultActMessage(actId, "SUCC");
     }
 
-    public long actId() {
+    public int actId() {
         return actId;
     }
 }
