+++ src/main/java/org/hashdb/ms/net/nio/msg/v1/DefaultActMessage.java
@@ -3,14 +3,14 @@ package org.hashdb.ms.net.nio.msg.v1;
 /**
  * Date: 2024/1/17 14:08
  *
- * @author huanyuMake-pecdle
+ * @author Huanyu Mark
  */
 public class DefaultActMessage extends ActMessage<String> {
-    public DefaultActMessage(long id, long actId, String body) {
+    public DefaultActMessage(int id, int actId, String body) {
         super(id, actId, body);
     }
 
-    public DefaultActMessage(long id, String body) {
+    public DefaultActMessage(int id, String body) {
         super(id, body);
     }
