+++ src/main/java/org/hashdb/ms/net/nio/msg/v1/ProtocolSwitchingMessage.java
@@ -6,10 +6,10 @@ import org.jetbrains.annotations.Nullable;
  * Date: 2024/1/27 16:11
  * 切换通讯协议
  *
- * @author huanyuMake-pecdle
+ * @author Huanyu Mark
  */
 public class ProtocolSwitchingMessage extends Message<Integer> {
-    public ProtocolSwitchingMessage(long id, @Nullable Integer protocolCode) {
+    public ProtocolSwitchingMessage(int id, @Nullable Integer protocolCode) {
         super(id, protocolCode);
     }
